Fixed time zone issues.
Users are now prevented from checking in if they have already checked into a party on a given date.

Homepage and dashboard now both show the correct list of parties that you can check into.

Replaced all instances of unix timestamps with DateTime objects.

Introduced a current_time() function that returns the curent GMT or local time. The rest of the code must make use of this. This allows us to fake the time for testing purposes.

Show doors have closed for checkin message if you can no longer checkin.